Hackaton15/mrodriguez/requirements.txt
asgiref==3.7.2
autopep8==2.0.2
black==23.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
Django==4.2.2
django-adminlte3==0.1.6
django-ckeditor==6.5.1
django-environ==0.10.0
django-filter==23.2
django-js-asset==2.0.0
djangorestframework==3.14.0
Flask==2.2.3
fonttools==4.39.2
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
mypy-extensions==1.0.0
mysql-connector-python==8.0.32
numpy==1.24.2
packaging==23.0
pathspec==0.11.0
Pillow==9.5.0
platformdirs==3.0.0
protobuf==3.20.3
psycopg2==2.9.6
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
Werkzeug==2.2.3